+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/config/object.php
@@ -9,6 +9,6 @@ return new class() {
     {
         $s = $c->services();
         $s->set(BarService::class)
-            ->args([inline('FooClass')]);
+            ->args([service('FooClass')]);
     }
 };